requirements.txt
cffi==1.15.1
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.18.0
pycryptodomex==3.18.0
PyNaCl==1.5.0
PySimpleGUI==4.60.5
